requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database
sqlalchemy==2.0.23
alembic==1.13.0
psycopg2-binary==2.9.9

# Task Queue
celery==5.3.4
redis==5.0.1

# Environment & Config
python-dotenv==1.0.0
pydantic-settings==2.1.0

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
isort==5.12.0

# Security
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2

# Email validation
email-validator==2.1.0

# Rate limiting
slowapi==0.1.9

# File validation
python-magic==0.4.27

# File processing
PyPDF2==3.0.1
python-docx==1.1.0
beautifulsoup4==4.12.2
lxml==4.9.4
aiofiles==23.2.1

# HTTP client
aiohttp==3.9.1

# Tavily API
tavily-python==0.6.0

# Google AI SDK
google-generativeai==0.8.3

